refactor: ターン相関の純関数を common/ へ（#1815 PR0）
サーバ側の runner が「送った文字列が相手の次の prompt に現れたターン」を
待つのに、その判断が src/composables/ にあって import できなかった。
両側が同じ判断をするので CLAUDE.md の common/ 規則どおり移す。

セルに出す文言（ExchangeOutcome / outcomeMessage）はブラウザだけのものなので
src に残す。再 export はせず、import 元 6 箇所を書き換えた。

ふるまいは変えていない。旧実装を verbatim にコピーしたハーネスで
17,145 比較・差分 0。変異検証も実施（PR 本文に内訳）。

refs #1815
